Temperature receiver Android Application
Code is not finished, it contains few bugs.
The application has features like setting a warning for the minimum and maximum temperature. For example, the user sets the maximum temperature warning to 25 degrees Celsius and closes the app. The user will then get an alert/notification if the temperature rises over 25 degrees.